requirements/versioned.txt
Django==4.2.5
djangorestframework==3.15.1
psycopg2-binary==2.9.5
debugpy
geopandas==0.14.3
pydevd-pycharm
drf-yasg==1.21.7